#2096 - Generify exporter

+++ inception/inception-curation/src/main/java/de/tudarmstadt/ukp/inception/curation/export/CuratedDocumentsExporter.java
@@ -18,7 +18,7 @@
 package de.tudarmstadt.ukp.inception.curation.export;
 
 import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.CURATION_USER;
-import static de.tudarmstadt.ukp.clarin.webanno.api.export.ProjectExportRequest.FORMAT_AUTO;
+import static de.tudarmstadt.ukp.clarin.webanno.api.export.FullProjectExportRequest.FORMAT_AUTO;
 import static de.tudarmstadt.ukp.clarin.webanno.model.Mode.CURATION;
 import static de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState.CURATION_FINISHED;
 import static de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState.CURATION_IN_PROGRESS;
@@ -49,8 +49,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import de.tudarmstadt.ukp.clarin.webanno.api.DocumentImportExportService;
 import de.tudarmstadt.ukp.clarin.webanno.api.DocumentService;
 import de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.CasStorageSession;
+import de.tudarmstadt.ukp.clarin.webanno.api.export.FullProjectExportRequest;
 import de.tudarmstadt.ukp.clarin.webanno.api.export.ProjectExportException;
-import de.tudarmstadt.ukp.clarin.webanno.api.export.ProjectExportRequest;
 import de.tudarmstadt.ukp.clarin.webanno.api.export.ProjectExportTaskMonitor;
 import de.tudarmstadt.ukp.clarin.webanno.api.export.ProjectExporter;
 import de.tudarmstadt.ukp.clarin.webanno.api.export.ProjectImportRequest;
@@ -101,7 +101,7 @@ public class CuratedDocumentsExporter
      *            The folder where curated documents are copied to be exported as Zip File
      */
     @Override
-    public void exportData(ProjectExportRequest aRequest, ProjectExportTaskMonitor aMonitor,
+    public void exportData(FullProjectExportRequest aRequest, ProjectExportTaskMonitor aMonitor,
             ExportedProject aExProject, File aStage)
         throws Exception
     {
@@ -138,34 +138,37 @@ public class CuratedDocumentsExporter
                         }
 
                         // Determine which format to use for export
-                        String formatId = FORMAT_AUTO.equals(aRequest.getFormat())
-                                ? sourceDocument.getFormat()
-                                : aRequest.getFormat();
+                        if (aRequest.getFormat() != null) {
+                            String formatId = FORMAT_AUTO.equals(aRequest.getFormat())
+                                    ? sourceDocument.getFormat()
+                                    : aRequest.getFormat();
 
-                        FormatSupport format = importExportService.getWritableFormatById(formatId)
-                                .orElseGet(() -> {
-                                    FormatSupport fallbackFormat = importExportService
-                                            .getFallbackFormat();
-                                    aMonitor.addMessage(LogMessage.warn(this,
-                                            "Curation: %s No writer found for original format [%s] "
-                                                    + "- exporting as [%s] instead.",
-                                            sourceDocument, formatId, fallbackFormat.getName()));
-                                    return fallbackFormat;
-                                });
+                            FormatSupport format = importExportService
+                                    .getWritableFormatById(formatId).orElseGet(() -> {
+                                        FormatSupport fallbackFormat = importExportService
+                                                .getFallbackFormat();
+                                        aMonitor.addMessage(LogMessage.warn(this,
+                                                "Curation: %s No writer found for original format [%s] "
+                                                        + "- exporting as [%s] instead.",
+                                                sourceDocument, formatId,
+                                                fallbackFormat.getName()));
+                                        return fallbackFormat;
+                                    });
 
-                        // Copy secondary export format for convenience - not used during import
-                        try {
-                            File curationFile = importExportService.exportAnnotationDocument(
-                                    sourceDocument, CURATION_USER, format, CURATION_USER, CURATION,
-                                    true, bulkOperationContext);
-                            copyFileToDirectory(curationFile, curationDir);
-                            forceDelete(curationFile);
-                        }
-                        catch (Exception e) {
-                            // error("Unexpected error while exporting project: " +
-                            // ExceptionUtils.getRootCauseMessage(e) );
-                            throw new ProjectExportException(
-                                    "Aborting due to unrecoverable error while exporting!");
+                            // Copy secondary export format for convenience - not used during import
+                            try {
+                                File curationFile = importExportService.exportAnnotationDocument(
+                                        sourceDocument, CURATION_USER, format, CURATION_USER,
+                                        CURATION, true, bulkOperationContext);
+                                copyFileToDirectory(curationFile, curationDir);
+                                forceDelete(curationFile);
+                            }
+                            catch (Exception e) {
+                                // error("Unexpected error while exporting project: " +
+                                // ExceptionUtils.getRootCauseMessage(e) );
+                                throw new ProjectExportException(
+                                        "Aborting due to unrecoverable error while exporting!");
+                            }
                         }
                     }
                 }
